refactor(mha): extract inner defs to module-level functions with ctx dict
Move 5 inner closures (apply_causal_mask, apply_kv_oob_mask,
fmha_pipeline_ctx, tile_iteration, _ep_finish) from fmha_fwd_kernel
to module-level functions in fmha_utils.py.

Captured kernel-scope variables are passed via a ctx dict built once
in the kernel prologue. This makes the kernel body purely sequential
(SECTION 1→2→3→4) with no interleaved function definitions.

fmha_kernel.py: 2090 → 1149 lines (kernel body is now pure control flow)
fmha_utils.py: 2990 → 3969 lines (gains the extracted pipeline functions)
